requirements.txt
paddlepaddle==2.4.2
paddleocr
streamlit
Pillow
torch
numpy